src/requirements.txt
torch>=1.9.0 --extra-index-url https://download.pytorch.org/whl/cu113
torchaudio>=0.9.0 --extra-index-url https://download.pytorch.org/whl/cu113
torchvision --extra-index-url https://download.pytorch.org/whl/cu113
torchsummary>=1.5.1
onnx
onnxruntime
torchvision
pytorch-crf
numpy>=1.19.5
scipy>=1.7.1
scikit-learn>=0.24.2
tqdm>=4.62.1
pyyaml>=5.3.1
HyperPyYAML
pydub
librosa>=0.8.1
SoundFile>=0.10.3.post1
matplotlib
Flask
Flask-RESTful
webrtcvad
nnAudio
autopep8
black
flake8
tensorboard
transformers
seqeval
folium
pandas
scipy
seaborn
04virtualenv -
jupyterlab
notebook
Pillow
opencv-python
requests
selenium
Scrapy3
openpyxl
pyaudio
asteroid
asteroid-filterbanks
torch-audiomentations
audiomentations
psutil
pytorch-metric-learning
simplejson
julius